requirements.txt
flask
chess
redis
python-dotenv